Use console event listener instead of overriding core command
